refactor x2
