+++ ts-client/src/dealExplorerClient/types/schemes.ts
